Moved factory to Stdlib namespace
